requirements.txt
transformers==4.31.0
numpy==1.22.4
gdown==4.6.4
Pillow==8.4.0